+++ src/components/planner/PlannerDetail.tsx
@@ -1,11 +1,10 @@
 "use client";
-import { PlannerParticipantRole, PlannerShareLink, ScheduleDetail } from "@/types/planner";
+import { PlannerParticipantRole, ScheduleDetail } from "@/types/planner";
 import PlannerTopHeader from "./top/PlannerTopHeader";
 import { useQuery } from "@tanstack/react-query";
 import { plannerQueries } from "@/queries/planner";
 import PlannerTopHeaderSkeleton from "../loading/planner/PlannerTopHeaderSkeleton";
 import PlannerTopActions from "./top/PlannerTopActions";
-import { getShareBaseUrl } from "@/utils/helpers/domain";
 import { concertQueries } from "@/queries/concerts";
 import PlannerBodySection from "./PlannerBodySection";
 import PlannerTimelineSectionSkeleton from "../loading/planner/PlannerTimelineSectionSkeleton";
@@ -13,19 +12,15 @@ import PlannerTopActionsSkeleton from "../loading/planner/PlannerTopActionsSkele
 
 interface PlannerDetailProps {
   planId: string;
-  role: PlannerParticipantRole;
+  userRole: PlannerParticipantRole;
   domain: string;
 }
 
-export default function PlannerDetail({ planId, role, domain }: PlannerDetailProps) {
+export default function PlannerDetail({ planId, userRole, domain }: PlannerDetailProps) {
   const { data: planDetail, isLoading } = useQuery(plannerQueries.detail(planId));
   const { data: concertDetail, isLoading: isConcertLoading } = useQuery(
     concertQueries.detail(planDetail?.concertId.toString() || "")
   );
-  const { data: shareData, isLoading: isShareLoading } = useQuery({
-    ...plannerQueries.share(planId),
-    enabled: role === "OWNER" || role === "EDITOR", // 공유 링크는 OWNER와 EDITOR만 필요하므로 조건부로 쿼리 실행
-  });
 
   if (isLoading || isConcertLoading || !planDetail || !concertDetail)
     return (
@@ -41,25 +36,10 @@ export default function PlannerDetail({ planId, role, domain }: PlannerDetailPro
     (schedule) => schedule.isMainEvent
   )!;
 
-  // 공유 링크 정보 구성
-  const shareLink: PlannerShareLink = { domain, url: "", status: "" };
-  if (shareData?.status === "ok") {
-    const baseUrl = getShareBaseUrl(domain);
-    shareLink.url = `${baseUrl}/planner/share?code=${shareData.data.shareToken}`;
-  } else if (isShareLoading) {
-    shareLink.status = "공유 링크를 불러오는 중입니다...";
-  } else if (shareData?.status === "not_created") {
-    shareLink.status = "아직 공유 링크가 생성되지 않았습니다.";
-  } else if (shareData?.status === "forbidden") {
-    shareLink.status = shareData.message;
-  } else if (shareData?.status === "error") {
-    shareLink.status = shareData.message;
-  }
-
   return (
     <>
-      <PlannerTopHeader planDetail={planDetail} role={role} concertDetail={concertDetail} />
-      {(role === "OWNER" || role === "EDITOR") && (
+      <PlannerTopHeader planDetail={planDetail} userRole={userRole} concertDetail={concertDetail} />
+      {(userRole === "OWNER" || userRole === "EDITOR") && (
         <PlannerTopActions
           concertCoords={{
             lat: concertSchedules.locationLat as number,
@@ -67,8 +47,8 @@ export default function PlannerDetail({ planId, role, domain }: PlannerDetailPro
           }}
           planId={planId}
           schedules={planDetail.schedules}
-          role={role}
-          shareLink={shareLink}
+          userRole={userRole}
+          domain={domain}
         />
       )}
       <PlannerBodySection
@@ -78,7 +58,7 @@ export default function PlannerDetail({ planId, role, domain }: PlannerDetailPro
           lat: concertSchedules.locationLat as number,
           lon: concertSchedules.locationLon as number,
         }}
-        role={role}
+        userRole={userRole}
         totalDuration={planDetail.totalDuration}
       />
     </>
